Created static footer
